refactor: try new folder structure, move a few files around
