+++ src/CoreNav.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
-import Icon from "./Icon";
+import Icon from "./components/Icon/Icon";
 import "./style/CoreNav.css";
 
 const CoreNav = ({ view, guides }) => (
