+++ src/components/Card/Card.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
-import "./style/Card.css";
+import "./Card.css";
 
 export default class Card extends Component {
   limitText = (text, charCount = 100) => {
